Simplify/PSR-2
Not quite sure what I was thinking when I originally built this. Deeply over-complexified. Simplified much of the logic to use built-in Smarty structures/methods instead of my own rickety, bolted-on logic. Also reformatted to be (almost entirely) PSR-2 compatible (BootstrapSmarty_Exception isn't a good class name by that standard, but it stays because, well, I refer to it elsewhere!).